+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   * Using a cache for updating THIDs of duplicated elements and their options.
   * Testing line duplication.
   * Created MPTHElementDuplicatorAux class.
+  * Make _updatedTHIDsMap an internal class variable on MPTHElementDuplicatorAux.
 
 ## 0.3.0 - 2026-03-09 - The [Great Fanfarra](https://youtu.be/HlV59UbrY-E) release
 * New features: